Fix bug #416 Random bookmarks
What is happening is that any article that is liked by any logged in user, has the users id added to the Bookmarks array, which is a property of the Post object. When prisma queries the DB, it selects the userId from Bookmarks where userId.
This seems to work fine when the user is logged in. However when userId is undefined it returns all the userId.
Fixed by checking if user is undefined when assigning the currentUserLikesPost variable.