fix: Footer didn't stick to the bottom.
I added an extra className "min-h-96" on the div
containing the post tabs.That made the footer goes
to the bottom of the page. I tried to make the
Footer get Sticked to the bottom, using an
"sticky bottom-0" on the Footer, but it didn't
work. The component must be flex.